Fix run script

+++ run.sh
@@ -2,10 +2,13 @@
 
 rm -rf allure-*
 
-sudo rm ~/.aerokube/selenoid/logs/*
-sudo rm ~/.aerokube/selenoid/video/*
+if [! -z "$(ls -A ~/.aerokube/selenoid/logs)"]; then
+    sudo rm ~/.aerokube/selenoid/logs/*
+fi
+if [! -z "$(ls -A ~/.aerokube/selenoid/video)"]; then
+    sudo rm ~/.aerokube/selenoid/video/*
+fi
 
 pytest -v --alluredir=allure-report
 
 allure serve allure-report
-
